pyproject.toml
[project]
name = "ncm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "siscomex-ncm>=1.0.2",
    "streamlit>=1.45.1",
]
